refactor(core): services/submissions.ts 1097 行 → 4 模块文件
PR-3: 拆分 submissions service 为单一职责的子模块，删除死代码 SubmissionWithResult。

## 拆分结果
- submissions-types.ts (122 行)：共享 DTO
  - SubmissionInput / SubmissionResponse / SubmissionDetail
  - SubmissionListItem / ListSubmissionsParams / ListSubmissionsResult
  - TodayStats
- submissions-crud.ts (481 行)：CRUD 核心
  - listSubmissions / createSubmission / getSubmission / deleteSubmission
  - 私有 parseDetails / MAX_OUTPUT_LENGTH
- submissions-result.ts (156 行)：评测结果写回
  - saveEvaluationResult（事务 + rejudge_seq 防护）
  - updateSubmissionStatus（状态机）
- submissions-rejudge.ts (326 行)：重测
  - rejudgeSubmission（单条）
  - rejudgeProblemSubmissions（批量，MAX_BATCH_REJUDGE=500）
- submissions-stats.ts (85 行)：统计
  - getTotalStats / getTodayStats
- submissions.ts (45 行)：barrel re-export
  - 保持 routes/admin.ts、routes/submissions.ts、routes/sse.ts、mq/consumer.ts
    既有 import 路径不变

## 死代码清理（PR-5 子任务）
- 删除 SubmissionWithResult 接口（全仓无引用）
- 删除 _toSubmissionResponse 私有函数（实际未被调用）

## 改进
- 单文件最大 481 行（原 1097 行），每个模块职责清晰
- 类型集中在 submissions-types.ts，CRUD/result/rejudge 互相通过 updateSubmissionStatus
  最小耦合（避免循环依赖）
- 测试覆盖无变化：tests/services/submissions.test.ts (13/13 通过) +
  tests/routes/submissions.test.ts (15/15 通过)

## 迁移指南
- 新代码：从对应子模块直接 import（如 import { getSubmission } from '../services/submissions-crud.ts'）
- 旧代码：通过 barrel 'submissions.ts' 继续可用（保持兼容）
